+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 2024-05-18
+
+### Fixed
+
+- Supply generation rules.
+- Jump punch damage.
+
 ## [2.0.1] - 2024-05-17
 
 ### Fixed
@@ -221,6 +228,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Nothing.
 
+[2.0.2]: https://github.com/thuasta/thuai-7/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/thuasta/thuai-7/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/thuasta/thuai-7/compare/v1.4.1...v2.0.0
 [1.4.1]: https://github.com/thuasta/thuai-7/compare/v1.4.0...v1.4.1